requirements.txt
python-dotenv
langchain
openai
youtube-transcript-api
pinecone-client
pgvector